config: Move bootloader code to its own file
